requirements.txt
torch==1.9.0
torchvision==0.10.0
segmentation-models-pytorch==0.2.1
opencv-python==4.5.5.64
numpy==1.19.5
Pillow==8.4.0
